+++ src/test/java/com/simisinc/platform/domain/model/DomainModelTest.java
@@ -14,29 +14,37 @@
  * limitations under the License.
  */
 
-package com.simisinc.platform.infrastructure.persistence;
+package com.simisinc.platform.domain.model;
 
 import org.junit.jupiter.api.Test;
 import org.meanbean.test.BeanVerifier;
 import org.meanbean.util.ClassPathUtils;
 
 import java.sql.Timestamp;
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * @author matt rajkowski
  * @created 5/11/2022 10:30 PM
  */
-class PersistenceTest {
-
+class DomainModelTest {
   @Test
-  void testSpecification() {
-    Class<?>[] beanClasses = ClassPathUtils.findClassesIn("com.simisinc.platform.infrastructure.persistence");
+  void testSettersAndGetters() {
+
+    // Verify complex classes separately
+    List<String> complexClasses = new ArrayList<>();
+
+    // ProductSku setter enforces changes for BeanUtils usage
+    complexClasses.add("com.simisinc.platform.domain.model.ecommerce.ProductSku");
+
+    // Test getters and setters for the domain model
+    Class<?>[] beanClasses = ClassPathUtils.findClassesIn("com.simisinc.platform.domain.model");
     for (Class k : beanClasses) {
       String thisClass = k.getName();
-      if (!thisClass.endsWith("Specification")) {
+      if (complexClasses.contains(thisClass)) {
         continue;
       }
-      // Specifications have getters and settings
       BeanVerifier.forClass(k)
           .editSettings()
           .registerFactory(Timestamp.class, () -> new Timestamp(System.currentTimeMillis()))
