+++ src/test/java/com/simisinc/platform/infrastructure/persistence/SpecificationTest.java
@@ -14,37 +14,29 @@
  * limitations under the License.
  */
 
-package com.simisinc.platform.domain;
+package com.simisinc.platform.infrastructure.persistence;
 
 import org.junit.jupiter.api.Test;
 import org.meanbean.test.BeanVerifier;
 import org.meanbean.util.ClassPathUtils;
 
 import java.sql.Timestamp;
-import java.util.ArrayList;
-import java.util.List;
 
 /**
  * @author matt rajkowski
  * @created 5/11/2022 10:30 PM
  */
-class DomainModelTest {
+class SpecificationTest {
+
   @Test
-  void testSettersAndGetters() {
-
-    // Verify complex classes separately
-    List<String> complexClasses = new ArrayList<>();
-
-    // ProductSku setter enforces changes for BeanUtils usage
-    complexClasses.add("com.simisinc.platform.domain.model.ecommerce.ProductSku");
-
-    // Test getters and setters for the domain model
-    Class<?>[] beanClasses = ClassPathUtils.findClassesIn("com.simisinc.platform.domain.model");
+  void test() {
+    Class<?>[] beanClasses = ClassPathUtils.findClassesIn("com.simisinc.platform.infrastructure.persistence");
     for (Class k : beanClasses) {
       String thisClass = k.getName();
-      if (complexClasses.contains(thisClass)) {
+      if (!thisClass.endsWith("Specification")) {
         continue;
       }
+      // Specifications have getters and settings
       BeanVerifier.forClass(k)
           .editSettings()
           .registerFactory(Timestamp.class, () -> new Timestamp(System.currentTimeMillis()))
